refactor: Move ctrlcreader.go under /k8s/internal/
CtrlCReader is only used by the k8s package.
