+++ k8s/internal/ctrlcreader.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package utils
+package internal
 
 import (
 	"context"
